+++ circus-train-core/src/main/java/com/hotels/bdp/circustrain/core/transformation/AbstractTableParametersTransformation.java
@@ -15,37 +15,49 @@
  */
 package com.hotels.bdp.circustrain.core.transformation;
 
+import static com.hotels.bdp.circustrain.core.conf.CircusTrainTransformOptions.TABLE_PROPERTIES;
+
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.Map;
 
+import com.hotels.bdp.circustrain.api.conf.TransformOptions;
 import com.hotels.bdp.circustrain.api.event.EventTableReplication;
 import com.hotels.bdp.circustrain.api.event.TableReplicationListener;
-import com.hotels.bdp.circustrain.core.conf.TableParametersConfig;
 
 public abstract class AbstractTableParametersTransformation implements TableReplicationListener {
 
-  private final TableParametersConfig tableParametersConfig;
-  private Map<String, String> tableParametersConfigOverride;
+  private final Map<String, String> tableParameters = new HashMap<>();
+  private Map<String, String> tableParametersOverride;
 
-  protected AbstractTableParametersTransformation(TableParametersConfig tableParametersConfig) {
-    this.tableParametersConfig = tableParametersConfig;
+  protected AbstractTableParametersTransformation(TransformOptions transformOptions) {
+    if (transformOptions.getTransformOptions() == null) {
+      return;
+    }
+    Object tableParameters = transformOptions.getTransformOptions().get(TABLE_PROPERTIES);
+    if (tableParameters instanceof Map) {
+      this.tableParameters.putAll((Map<String, String>) tableParameters);
+    }
   }
 
   protected Map<String, String> getTableParameters() {
-    if (tableParametersConfigOverride != null && !tableParametersConfigOverride.isEmpty()) {
-      return tableParametersConfigOverride;
+    if (tableParametersOverride != null && !tableParametersOverride.isEmpty()) {
+      return tableParametersOverride;
     }
-    return tableParametersConfig.getParameters();
+    return tableParameters;
   }
 
   @SuppressWarnings("unchecked")
   @Override
   public void tableReplicationStart(EventTableReplication tableReplication, String eventId) {
-    tableParametersConfigOverride = Collections.emptyMap();
+    tableParametersOverride = Collections.emptyMap();
     Map<String, Object> transformOptions = tableReplication.getTransformOptions();
-    Object tableParametersOverride = transformOptions.get(TableParametersConfig.TABLE_REPLICATION_TABLE_PARAMETERS);
-    if (tableParametersOverride != null && tableParametersOverride instanceof Map) {
-      tableParametersConfigOverride = (Map<String, String>) tableParametersOverride;
+    if (transformOptions == null) {
+      return;
+    }
+    Object tableParametersOverride = transformOptions.get(TABLE_PROPERTIES);
+    if (tableParametersOverride instanceof Map) {
+      this.tableParametersOverride = (Map<String, String>) tableParametersOverride;
     }
   }